refactor: product/auth 쿼리키 별도 파일로 분리
