+++ src/features/auth/queryKeys.ts
@@ -1,0 +1,4 @@
+export const authKeys = {
+  all: ['auth'] as const,
+  user: () => [...authKeys.all, 'user'] as const,
+};
